scalability framework: move paths file
