+++ misc/python/materialize/scalability/io/paths.py
